+++ ITTWEB-ASPNetCore/Core/Domain/Component.cs
@@ -5,14 +5,14 @@ namespace ITTWEB_ASPNetCore.Core.Domain
 {
     public class Component
     {
-        public long ComponentId { get; set; }
-        public long ComponentTypeId { get; set; }
+        public int ComponentId { get; set; }
+        public int ComponentTypeId { get; set; }
         public int ComponentNumber { get; set; }
         public string SerialNo { get; set; }
         public ComponentStatus Status { get; set; }
         public string AdminComment { get; set; }
         public string UserComment { get; set; }
-        public long? CurrentLoanInformationId { get; set; }
+        public int? CurrentLoanInformationId { get; set; }
     }
 
     public static class ComponentMock
